+++ pypermedia/gzip_requests.py
@@ -3,8 +3,6 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-__author__ = 'alexmaskovyak'
-
 from requests import Request
 import zlib